packages/server/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cua-house-server"
version = "0.1.0"
description = "Computer-use VM sandbox orchestration server"
requires-python = ">=3.12,<3.14"
license = { text = "MIT" }
dependencies = [
    "cua-house-common",
    "fastapi>=0.115.0",
    "httpx>=0.27.0",
    "pydantic>=2.8.0",
    "PyYAML>=6.0",
    "uvicorn>=0.27.0",
    "websockets>=12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=1.3.0",
]

[project.scripts]
cua-house-server = "cua_house_server.cli:main"

[tool.uv.sources]
cua-house-common = { workspace = true }

[tool.hatch.build.targets.wheel]
packages = ["src/cua_house_server"]
